+++ web/src/components/inputs/NumberInput.tsx
@@ -7,10 +7,55 @@ import EditableInput from "./EditableInput";
 import { useFocusPan } from "../../hooks/useFocusPan";
 import { useViewport } from "@xyflow/react";
 import { getMousePosition } from "../../utils/MousePosition";
-import { InputProps, NumberInputState } from "./NumberInput.types";
-import { useDragHandling } from "./NumberInput.hooks";
+import { useDragHandling } from "../../hooks/useNumberInput";
 import DisplayValue from "./DisplayValue";
 import SpeedDisplay from "./SpeedDisplay";
+
+// Debug mode
+export const DEBUG = false;
+
+// Drag-tuning constants
+export const DRAG_THRESHOLD = 10; // px before drag counts
+export const DRAG_SLOWDOWN_DEAD_ZONE_PX = 10; // no-slow band above/below slider
+export const DRAG_SLOWDOWN_RAMP_BASE_PX = 450; // ramp half-height in px
+export const DRAG_SLOWDOWN_RAMP_PX = DRAG_SLOWDOWN_RAMP_BASE_PX * 2; // full ramp span
+export const MIN_SPEED_FACTOR = 0.1; // min speed (normal)
+export const SHIFT_MIN_SPEED_FACTOR = 0.01; // min speed with Shift
+export const SHIFT_SLOWDOWN_DIVIDER = 5; // Shift divides speed by this
+export const REFERENCE_SLIDER_WIDTH = 180; // px - base slider width for zoom calculations
+
+export interface InputProps {
+  nodeId: string;
+  name: string;
+  description?: string | null;
+  min?: number;
+  max?: number;
+  value: number;
+  onChange: (event: any, value: number) => void;
+  id: string;
+  size?: "small" | "medium";
+  color?: "primary" | "secondary";
+  className?: string;
+  inputType?: "int" | "float";
+  hideLabel?: boolean;
+  tabIndex?: number;
+  zoomAffectsDragging?: boolean;
+}
+
+export interface NumberInputState {
+  isDefault: boolean;
+  localValue: string;
+  originalValue: number;
+  dragStartX: number;
+  decimalPlaces: number;
+  isDragging: boolean;
+  hasExceededDragThreshold: boolean;
+  dragInitialValue: number;
+  currentDragValue: number;
+  lastClientX: number;
+  dragPointerY: number | null;
+  actualSliderWidth: number;
+}
 
 const NumberInput: React.FC<InputProps> = (props) => {
   const [state, setState] = useState<NumberInputState>({
